made models an independant project
