+++ README.md
@@ -11,4 +11,10 @@ For any IDE, two environment variables are needed:
 * DefaultConnection: `[database connection string]`
 * TokenAuthentication_SecretKey: `[any secret string]`
 
-the variables have been provided.
+The variables have been provided.
+
+migrations is used to update the database.
+for more info: http://benjii.me/2017/05/enable-entity-framework-core-migrations-visual-studio-2017/
+make sure you're in `src` and also make sure envrionment variable `DefaultConnection` is set
+Note: I have not been able to get the Package Manager Console to work yet.
+
