+++ models/FrameworksAndLibs.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace portfolio_api.Models
+namespace PortfolioApi.Models
 {
     public class FrameworksAndLibs : RankableItem
     {
