+++ models/Phones/PhoneNumber.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace portfolio_api.Models.Phones
+namespace PortfolioApi.Models.Phones
 {
     public class PhoneNumber
     {
