+++ src/Controllers/ContactController.cs
@@ -2,10 +2,10 @@ using System.Linq;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using portfolio_api.Services;
-using Model = portfolio_api.Models.Contact;
+using PortfolioApi.Services;
+using Model = PortfolioApi.Models.Contact;
 
-namespace portfolio_api.Controllers
+namespace PortfolioApi.Controllers
 {
     [Route("api/[controller]")]
     public class ContactController: PortfolioController
