+++ src/PortfolioApi.csproj
@@ -19,4 +19,7 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="1.1.2" />
     <DotNetCliToolReference Include="Microsoft.EntityFrameworkCore.Tools.DotNet" Version="1.0.0" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\models\PortfolioApi.Models.csproj" />
+  </ItemGroup>
 </Project>
