--- test/UnitTest1.cs
@@ -1,16 +1,0 @@
-using System;
-using Xunit;
-
-namespace test
-{
-    public class UnitTest1
-    {
-        // https://docs.microsoft.com/en-us/dotnet/articles/core/testing/unit-testing-with-dotnet-test
-        
-        [Fact]
-        public void Test1()
-        {
-            // No Tests yet
-        }
-    }
-}
