+++ tests/UnitTest1.cs
@@ -1,0 +1,13 @@
+using System;
+using Xunit;
+
+namespace PortfolioApi.Tests
+{
+    public class UnitTest1
+    {
+        [Fact]
+        public void Test1()
+        {
+        }
+    }
+}
